Fix code duplication issue.
